refactor: 디렉토리 재구성 및 interface를 통해 구현체를 갈아 끼울 수 있도록 진행
