+++ src/main/java/kitchenpos/infrastructure/jdbc/JdbcTemplateOrderDao.java
@@ -1,5 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.infrastructure.jdbc;
 
+import kitchenpos.port.OrderPort;
 import kitchenpos.domain.Order;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
@@ -18,7 +19,7 @@ import java.util.Objects;
 import java.util.Optional;
 
 @Repository
-public class JdbcTemplateOrderDao implements OrderDao {
+public class JdbcTemplateOrderDao {
     private static final String TABLE_NAME = "orders";
     private static final String KEY_COLUMN_NAME = "id";
 
@@ -33,7 +34,6 @@ public class JdbcTemplateOrderDao implements OrderDao {
         ;
     }
 
-    @Override
     public Order save(final Order entity) {
         if (Objects.isNull(entity.getId())) {
             final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
@@ -44,7 +44,6 @@ public class JdbcTemplateOrderDao implements OrderDao {
         return entity;
     }
 
-    @Override
     public Optional<Order> findById(final Long id) {
         try {
             return Optional.of(select(id));
@@ -53,13 +52,11 @@ public class JdbcTemplateOrderDao implements OrderDao {
         }
     }
 
-    @Override
     public List<Order> findAll() {
         final String sql = "SELECT id, order_table_id, order_status, ordered_time FROM orders";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    @Override
     public boolean existsByOrderTableIdAndOrderStatusIn(final Long orderTableId, final List<String> orderStatuses) {
         final String sql = "SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END" +
                 " FROM orders WHERE order_table_id = (:orderTableId) AND order_status IN (:orderStatuses)";
@@ -69,7 +66,6 @@ public class JdbcTemplateOrderDao implements OrderDao {
         return jdbcTemplate.queryForObject(sql, parameters, Boolean.class);
     }
 
-    @Override
     public boolean existsByOrderTableIdInAndOrderStatusIn(final List<Long> orderTableIds, final List<String> orderStatuses) {
         final String sql = "SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END" +
                 " FROM orders WHERE order_table_id IN (:orderTableIds) AND order_status IN (:orderStatuses)";
