+++ src/main/java/kitchenpos/infrastructure/jpa/adapter/OrderJpaRepository.java
@@ -1,0 +1,37 @@
+package kitchenpos.infrastructure.jpa.adapter;
+
+import kitchenpos.domain.Order;
+import kitchenpos.port.OrderPort;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public class OrderJpaRepository implements OrderPort {
+    @Override
+    public Order save(Order entity) {
+        return null;
+    }
+
+    @Override
+    public Optional<Order> findById(Long id) {
+        return Optional.empty();
+    }
+
+    @Override
+    public List<Order> findAll() {
+        return null;
+    }
+
+    @Override
+    public boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<String> orderStatuses) {
+        return false;
+    }
+
+    @Override
+    public boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<String> orderStatuses) {
+        return false;
+    }
+}
