+++ src/main/java/kitchenpos/infrastructure/jpa/repository/MenuJpaRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.infrastructure.jpa.repository;
+
+import kitchenpos.domain.menu.Menu;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MenuJpaRepository extends JpaRepository<Menu, Long> {
+}
