+++ src/main/java/kitchenpos/infrastructure/jpa/repository/OrderJpaRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.infrastructure.jpa.repository;
+
+import kitchenpos.domain.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderJpaRepository extends JpaRepository<Order, Long> {
+}
