+++ src/main/java/kitchenpos/infrastructure/jpa/repository/ProductJpaRepository.java
@@ -1,0 +1,13 @@
+package kitchenpos.infrastructure.jpa.repository;
+
+import kitchenpos.domain.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ProductJpaRepository extends JpaRepository<Product, Long> {
+
+    List<Product> findAllByIdIn(List<Long> id);
+}
