+++ src/main/java/kitchenpos/infrastructure/jpa/repository/TableGroupJpaRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.infrastructure.jpa.repository;
+
+import kitchenpos.domain.TableGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TableGroupJpaRepository extends JpaRepository<TableGroup, Long> {
+}
