+++ src/main/java/kitchenpos/port/MenuGroupPort.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao;
+package kitchenpos.port;
 
 import kitchenpos.domain.MenuGroup;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuGroupDao {
+public interface MenuGroupPort {
     MenuGroup save(MenuGroup entity);
 
     Optional<MenuGroup> findById(Long id);
