+++ src/main/java/kitchenpos/port/MenuProductPort.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao;
+package kitchenpos.port;
 
 import kitchenpos.domain.MenuProduct;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuProductDao {
+public interface MenuProductPort {
     MenuProduct save(MenuProduct entity);
 
     Optional<MenuProduct> findById(Long id);
