+++ src/main/java/kitchenpos/port/OrderLineItemPort.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao;
+package kitchenpos.port;
 
 import kitchenpos.domain.OrderLineItem;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderLineItemDao {
+public interface OrderLineItemPort {
     OrderLineItem save(OrderLineItem entity);
 
     Optional<OrderLineItem> findById(Long id);
