+++ src/main/java/kitchenpos/port/OrderTablePort.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao;
+package kitchenpos.port;
 
 import kitchenpos.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderTableDao {
+public interface OrderTablePort {
     OrderTable save(OrderTable entity);
 
     Optional<OrderTable> findById(Long id);
