+++ src/main/java/kitchenpos/port/ProductPort.java
@@ -1,14 +1,16 @@
-package kitchenpos.dao;
+package kitchenpos.port;
 
 import kitchenpos.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductDao {
+public interface ProductPort {
     Product save(Product entity);
 
     Optional<Product> findById(Long id);
 
     List<Product> findAll();
+
+    List<Product> findAllByIdIn(List<Long> productId);
 }
