+++ src/main/java/kitchenpos/port/TableGroupPort.java
@@ -1,11 +1,11 @@
-package kitchenpos.dao;
+package kitchenpos.port;
 
 import kitchenpos.domain.TableGroup;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface TableGroupDao {
+public interface TableGroupPort {
     TableGroup save(TableGroup entity);
 
     Optional<TableGroup> findById(Long id);
